factor out assertions from rule analysis
